requirements.txt
PyQt6>=6.6.0
pyautogui>=0.9.54
Pillow>=10.0.0
img2pdf>=0.5.1
pyobjc-framework-Quartz>=10.0
pyobjc-framework-Cocoa>=10.0
pypdf>=4.0.0
